file out

+++ install_dependencies.sh
@@ -9,13 +9,17 @@ fi
 mkdir -p lib/
 cd lib/
 
-if ! [ -x liblogicalaccess/ ]; then
-  git clone https://github.com/islog/liblogicalaccess.git
-  cd liblogicalaccess
+if [ "$1" != "-nopull" ]; then
+  if ! [ -x liblogicalaccess/ ]; then
+    git clone https://github.com/islog/liblogicalaccess.git
+    cd liblogicalaccess
+  else
+    cd liblogicalaccess
+    git fetch origin
+    git reset --hard origin/master
+  fi
 else
   cd liblogicalaccess
-  git fetch origin
-  git reset --hard origin/master
 fi
 
 mkdir -p install/
@@ -23,5 +27,5 @@ cd install/
 cmake ../
 make install
 cd /usr/local/include/logicalaccess/
-grep --include=*.hpp -rnl './' -e "def UNIX" | xargs -i@ sed -i 's/def UNIX/def linux/g' @
-grep --include=*.hpp -rnl './' -e "defined(UNIX)" | xargs -i@ sed -i 's/defined(UNIX)/defined(linux)/g' @
+grep --include=*.hpp -rnl './' -e "def UNIX" | xargs -d \n -i@ sed -i 's/def UNIX/def linux/g' @
+grep --include=*.hpp -rnl './' -e "defined(UNIX)" | xargs -d \n -i@ sed -i 's/defined(UNIX)/defined(linux)/g' @
